+++ packages/svelte-undp-components/src/lib/components/index.ts
@@ -2,6 +2,7 @@ import Accordion from './Accordion.svelte';
 import BackToTop from './BackToTop.svelte';
 import Breadcrumbs, { type BreadcrumbPage } from './Breadcrumbs.svelte';
 import ColorMapPicker from './ColorMapPicker.svelte';
+import DatePicker from './DatePicker.svelte';
 import FieldControl from './FieldControl.svelte';
 import FloatingPanel from './FloatingPanel.svelte';
 import Help from './Help.svelte';
@@ -22,6 +23,7 @@ export {
 	BackToTop,
 	Breadcrumbs,
 	ColorMapPicker,
+	DatePicker,
 	FieldControl,
 	FloatingPanel,
 	Help,