+++ sites/geohub/src/routes/(app)/management/stac/catalog/[id]/+page.ts
@@ -1,7 +1,7 @@
 import type { PageLoad } from './$types';
 
 export const load: PageLoad = async ({ data }) => {
-	const { stac, datasetId, isRegistered } = data;
+	const { stac, datasetId, isRegistered, dataset, titilerUrl } = data;
 	const title = `${stac.name} | STAC Catalog management | GeoHub`;
 	const content = stac.name;
 
@@ -10,6 +10,8 @@ export const load: PageLoad = async ({ data }) => {
 		content,
 		stac,
 		datasetId,
-		isRegistered
+		isRegistered,
+		dataset,
+		titilerUrl
 	};
 };